chore(license): Relicense library + service to MPL-2.0
As the sole author of the library and service, I elect to relicense the
code that I have written from GPL-3.0 to MPL-2.0. This is to preserve
copyleft-ability while permitting linking of pop-launcher's libraries.

The launcher binary and its plugins shall remain as GPL-3.0, as they are
separate binaries with no need for linking to a frontend.