fix(desktop_entries): Respect `NoDisplay` precedence
Closes: pop-os/cosmic-epoch#905

`NoDisplay` allows applications or users to signal that an entry
shouldn't be shown in menus. This is useful for entries that may launch
an app in a certain way such as opening a specific menu or handling a
MIME type. Some desktop environments use it to signal special cases.

`NoDisplay` as well as `Hidden` allow user defined entries to hide apps.
User defined desktop entries should override the system's applications,
but the old logic continued parsing entries from all locations instead
of stopping at a user override.