fix: use /usr/libexec/.. for plugins instead of /usr/lib/..
This commit changes plugins to use /usr/libexec instead of /usr/lib. Since the plugins contain executables, the more apt directory for this need is /usr/libexec. Fedora is requiring this in order to package pop-launcher and cosmic-launcher, so this patch will be used for downstream Fedora, but if upstream would like to adopt this change, I've brought it here just in case! :)

Signed-off-by: Ryan Brue <ryanbrue.dev@gmail.com>