desktop_entries: Reload .desktop on search
Since pop-launcher has been demonized, the desktop_entries plugin isn't
relaunched on every search anymore. In consequence, it is not providing
up-to-date results when .desktop files have changed (e.g., on
un-/installation of a new application) anymore.

Move the app.reload() call into the request loop to trigger a re-index.
In a release-build this has been fine performance wise on an 8th Gen i5
laptop device. As a small performance improvement, only trigger this update
on a search event and if the search query is empty.

Alternatives could include remembering the last time the index was built
(initialised to UNIX_EPOCH before the loop) and only conditionally
triggering a re-index after a hot-period of 10s since the last run. In
our testing this didn't result in any noticeable difference.